fix flannel configuration
On master the default network interface is wlan0. On all nodes, it's
eth0. Therefore, master can't communicate with the nodes and pods don't
have any connectivity.

This patch changes the default interface to whatever one matches
`10.0.0.x`, which is the IP range for the internal network, so nodes
have access to internet.